support replace folder for overlayfs

+++ customize.sh
@@ -1,6 +1,10 @@
-SKIPUNZIP=1
-echo "- Perserve all changes..."
 MODID=$(basename "$MODPATH")
+
+if [ -d "/data/adb/modules/$MODID" ]; then
+echo "- Perserve all changes..."
+SKIPUNZIP=1
+mkdir -p "/data/adb/modules/$MODID"
 unzip -o "$ZIPFILE" 'post-fs-data.sh' 'module.prop' -d "/data/adb/modules/$MODID"
 rm -rf "$MODPATH"
 (sleep 1; rm -rf "/data/adb/modules/$MODID/update")&
+fi
